timex_span v0.0.1 TimexSpan

Types

t :: %TimexSpan{days: term, hours: term, mins: term, secs: term, weeks: term, years: term}

Functions

ago(span)

Specs

ago(t) :: Timex.DateTime.t

Return a DateTime that is earlier than the current datetime by the given span. Equivalent to calling before(Timex.Date.now)

before(span, date)

Specs

before(t, date :: Timex.DateTime.t) :: Timex.DateTime.t

Return a DateTime that is earlier than the given datetime by the given span.

Examples

iex> span = TimexSpan.years(5)
iex> date = Timex.Date.from({{2015, 6, 24}, {14, 27, 52}})
iex> TimexSpan.before(span, date) |> Timex.DateFormat.format!("{RFC1123}")
"Thu, 24 Jun 2010 14:27:52 +0000"

iex> import TimexSpan
iex> date = Timex.Date.from({{2015, 6, 24}, {14, 27, 52}})
iex> 5 |> years |> before(date) |> Timex.DateFormat.format!("{RFC1123}")
"Thu, 24 Jun 2010 14:27:52 +0000"

days(num_days, span \\ %TimexSpan{})

Specs

days(number, t) :: t

Create a span of days equal to the specified number. Optionally adds or replaces the day component of a span passed in as its second argument.

Examples

iex> TimexSpan.days(4)
%TimexSpan{days: 4}

iex> span = %TimexSpan{days: 3, secs: 2}
iex> TimexSpan.days(4, span)
%TimexSpan{days: 4, secs: 2}

from(span, date)

Specs

from(t, date :: Timex.DateTime.t) :: Timex.DateTime.t

Return a DateTime that is later than the given datetime by the given span.

Examples

iex> span = TimexSpan.years(5)
iex> date = Timex.Date.from({{2015, 6, 24}, {14, 27, 52}})
iex> TimexSpan.from(span, date) |> Timex.DateFormat.format!("{RFC1123}")
"Wed, 24 Jun 2020 14:27:52 +0000"

iex> import TimexSpan
iex> date = Timex.Date.from({{2015, 6, 24}, {14, 27, 52}})
iex> 5 |> years |> from(date) |> Timex.DateFormat.format!("{RFC1123}")
"Wed, 24 Jun 2020 14:27:52 +0000"

hours(num_hours, span \\ %TimexSpan{})

Specs

hours(number, t) :: t

Create a span of hours equal to the specified number. Optionally adds or replaces the hour component of a span passed in as its second argument.

Examples

iex> TimexSpan.hours(4)
%TimexSpan{hours: 4}

iex> span = %TimexSpan{hours: 3, secs: 2}
iex> TimexSpan.hours(4, span)
%TimexSpan{hours: 4, secs: 2}

mins(num_mins, span \\ %TimexSpan{})

Specs

mins(number, t) :: t

Create a span of mins equal to the specified number. Optionally adds or replaces the min component of a span passed in as its second argument.

Examples

iex> TimexSpan.mins(4)
%TimexSpan{mins: 4}

iex> span = %TimexSpan{mins: 3, secs: 2}
iex> TimexSpan.mins(4, span)
%TimexSpan{mins: 4, secs: 2}

secs(num_secs, span \\ %TimexSpan{})

Specs

secs(number, t) :: t

Create a span of secs equal to the specified number. Optionally adds or replaces the sec component of a span passed in as its second argument.

Examples

iex> TimexSpan.secs(4)
%TimexSpan{secs: 4}

iex> span = %TimexSpan{hours: 1, secs: 3}
iex> TimexSpan.secs(4, span)
%TimexSpan{hours: 1, secs: 4}

weeks(num_weeks, span \\ %TimexSpan{})

Specs

weeks(number, t) :: t

Create a span of weeks equal to the specified number. Optionally adds or replaces the week component of a span passed in as its second argument.

Examples

iex> TimexSpan.weeks(4)
%TimexSpan{weeks: 4}

iex> span = %TimexSpan{weeks: 3, secs: 2}
iex> TimexSpan.weeks(4, span)
%TimexSpan{weeks: 4, secs: 2}

years(num_years, span \\ %TimexSpan{})

Specs

years(number, t) :: t

Create a span of years equal to the specified number. Optionally adds or replaces the year component of a span passed in as its second argument.

Examples

iex> TimexSpan.years(4)
%TimexSpan{years: 4}

iex> span = %TimexSpan{years: 3, weeks: 2}
iex> TimexSpan.years(4, span)
%TimexSpan{years: 4, weeks: 2}