timex_sugar v0.0.1 TimexSugar

Functions

ago(span)

Specs

ago(span :: TimexSugar.Span.t) :: Timex.DateTime.t

Return a DateTime that is earlier than the current datetime by the given span. Equivalent to calling before(Timex.Date.now)

before(span, date)

Specs

before(span :: TimexSugar.Span.t, date :: Timex.DateTime.t) :: Timex.DateTime.t

Return a DateTime that is earlier than the given datetime by the given span.

Examples

iex> span = Elixir.TimexSugar.years(5)
iex> date = Timex.Date.from({{2015, 6, 24}, {14, 27, 52}})
iex> Elixir.TimexSugar.before(span, date)
%Timex.DateTime{calendar: :gregorian, day: 24, hour: 14, minute: 27, month: 6,
 ms: 0, second: 52,
 timezone: %Timex.TimezoneInfo{abbreviation: "UTC", from: :min,
  full_name: "UTC", offset_std: 0, offset_utc: 0, until: :max}, year: 2010}

iex> import TimexSugar
iex> date = Timex.Date.from({{2015, 6, 24}, {14, 27, 52}})
iex> 5 |> years |> before(date)
%Timex.DateTime{calendar: :gregorian, day: 24, hour: 14, minute: 27, month: 6,
 ms: 0, second: 52,
 timezone: %Timex.TimezoneInfo{abbreviation: "UTC", from: :min,
  full_name: "UTC", offset_std: 0, offset_utc: 0, until: :max}, year: 2010}

days(num_days, span \\ %TimexSugar.Span{})

Specs

days(number, span :: TimexSugar.Span.t) :: TimexSugar.Span.t

Create a span of days equal to the specified number. Optionally adds or replaces the day component of a span passed in as its second argument.

Examples

iex> Elixir.TimexSugar.days(4)
%Elixir.TimexSugar.Span{days: 4}

iex> span = %Elixir.TimexSugar.Span{days: 3, secs: 2}
iex> Elixir.TimexSugar.days(4, span)
%Elixir.TimexSugar.Span{days: 4, secs: 2}

from(span, date)

Specs

from(span :: TimexSugar.Span.t, date :: Timex.DateTime.t) :: Timex.DateTime.t

Return a DateTime that is later than the given datetime by the given span.

Examples

iex> span = Elixir.TimexSugar.years(5)
iex> date = Timex.Date.from({{2015, 6, 24}, {14, 27, 52}})
iex> Elixir.TimexSugar.from(span, date)
%Timex.DateTime{calendar: :gregorian, day: 24, hour: 14, minute: 27, month: 6,
 ms: 0, second: 52,
 timezone: %Timex.TimezoneInfo{abbreviation: "UTC", from: :min,
  full_name: "UTC", offset_std: 0, offset_utc: 0, until: :max}, year: 2020}

iex> import TimexSugar
iex> date = Timex.Date.from({{2015, 6, 24}, {14, 27, 52}})
iex> 5 |> years |> from(date)
%Timex.DateTime{calendar: :gregorian, day: 24, hour: 14, minute: 27, month: 6,
 ms: 0, second: 52,
 timezone: %Timex.TimezoneInfo{abbreviation: "UTC", from: :min,
  full_name: "UTC", offset_std: 0, offset_utc: 0, until: :max}, year: 2020}

hours(num_hours, span \\ %TimexSugar.Span{})

Specs

hours(number, span :: TimexSugar.Span.t) :: TimexSugar.Span.t

Create a span of hours equal to the specified number. Optionally adds or replaces the hour component of a span passed in as its second argument.

Examples

iex> Elixir.TimexSugar.hours(4)
%Elixir.TimexSugar.Span{hours: 4}

iex> span = %Elixir.TimexSugar.Span{hours: 3, secs: 2}
iex> Elixir.TimexSugar.hours(4, span)
%Elixir.TimexSugar.Span{hours: 4, secs: 2}

mins(num_mins, span \\ %TimexSugar.Span{})

Specs

mins(number, span :: TimexSugar.Span.t) :: TimexSugar.Span.t

Create a span of mins equal to the specified number. Optionally adds or replaces the min component of a span passed in as its second argument.

Examples

iex> Elixir.TimexSugar.mins(4)
%Elixir.TimexSugar.Span{mins: 4}

iex> span = %Elixir.TimexSugar.Span{mins: 3, secs: 2}
iex> Elixir.TimexSugar.mins(4, span)
%Elixir.TimexSugar.Span{mins: 4, secs: 2}

secs(num_secs, span \\ %TimexSugar.Span{})

Specs

secs(number, span :: TimexSugar.Span.t) :: TimexSugar.Span.t

Create a span of secs equal to the specified number. Optionally adds or replaces the sec component of a span passed in as its second argument.

Examples

iex> Elixir.TimexSugar.secs(4)
%Elixir.TimexSugar.Span{secs: 4}

iex> span = %Elixir.TimexSugar.Span{hours: 1, secs: 3}
iex> Elixir.TimexSugar.secs(4, span)
%Elixir.TimexSugar.Span{hours: 1, secs: 4}

weeks(num_weeks, span \\ %TimexSugar.Span{})

Specs

weeks(number, span :: TimexSugar.Span.t) :: TimexSugar.Span.t

Create a span of weeks equal to the specified number. Optionally adds or replaces the week component of a span passed in as its second argument.

Examples

iex> Elixir.TimexSugar.weeks(4)
%Elixir.TimexSugar.Span{weeks: 4}

iex> span = %Elixir.TimexSugar.Span{weeks: 3, secs: 2}
iex> Elixir.TimexSugar.weeks(4, span)
%Elixir.TimexSugar.Span{weeks: 4, secs: 2}

years(num_years, span \\ %TimexSugar.Span{})

Specs

years(number, span :: TimexSugar.Span.t) :: TimexSugar.Span.t

Create a span of years equal to the specified number. Optionally adds or replaces the year component of a span passed in as its second argument.

Examples

iex> Elixir.TimexSugar.years(4)
%Elixir.TimexSugar.Span{years: 4}

iex> span = %Elixir.TimexSugar.Span{years: 3, weeks: 2}
iex> Elixir.TimexSugar.years(4, span)
%Elixir.TimexSugar.Span{years: 4, weeks: 2}